final qualifier 2 changes, question mark
