+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/notquali2/close_red_9.java
@@ -1,17 +1,14 @@
-package org.firstinspires.ftc.teamcode.auto;
+package org.firstinspires.ftc.teamcode.auto.notquali2;
 
 import androidx.annotation.NonNull;
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.dashboard.telemetry.TelemetryPacket;
 import com.acmerobotics.roadrunner.Action;
-import com.acmerobotics.roadrunner.AngularVelConstraint;
 import com.acmerobotics.roadrunner.ParallelAction;
 import com.acmerobotics.roadrunner.Pose2d;
-import com.acmerobotics.roadrunner.ProfileAccelConstraint;
 import com.acmerobotics.roadrunner.SequentialAction;
 import com.acmerobotics.roadrunner.TrajectoryActionBuilder;
 import com.acmerobotics.roadrunner.TranslationalVelConstraint;
-import com.acmerobotics.roadrunner.TurnConstraints;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
@@ -24,14 +21,16 @@ import com.qualcomm.robotcore.hardware.VoltageSensor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 import org.firstinspires.ftc.teamcode.MecanumDrive;
 import org.firstinspires.ftc.teamcode.shooter.ShooterControl;
+import org.firstinspires.ftc.teamcode.tuning.shooter.RobotConstants;import org.firstinspires.ftc.teamcode.classes.Transfer_Values;
+
+
 @Config
-@Autonomous(name = "far blue 9 no mechanisms", group = "Autonomous")
-//@Disabled
+@Autonomous(name = "close red 9", group = "Autonomous")
+@Disabled
 //psuedocode
 /*
-
  */
-public class far_blue_9_no_mecha extends LinearOpMode{
+public class close_red_9 extends LinearOpMode{
 
     // if odometry is not properly tuned or constantly being retuned:
     // you MIGHT find it useful to change these values and use multiples of them instead of direct number
@@ -46,6 +45,7 @@ public class far_blue_9_no_mecha extends LinearOpMode{
     private ShooterControl flywheel;
 
     VoltageSensor battery;
+    private Transfer_Values transferValues;
 
 
 
@@ -227,6 +227,10 @@ public class far_blue_9_no_mecha extends LinearOpMode{
         public class Hold implements Action {
             @Override
             public boolean run(@NonNull TelemetryPacket packet) {
+                flywheel.setKf(RobotConstants.kF);
+                flywheel.setKp(RobotConstants.kP);
+                flywheel.setKi(RobotConstants.kI);
+                flywheel.setKd(RobotConstants.kD);
                 flywheel.setTargetRPM(940);
                 return false; // true reruns action
             }
@@ -256,67 +260,33 @@ public class far_blue_9_no_mecha extends LinearOpMode{
 
         flywheel = new ShooterControl(hardwareMap);
 
-        Pose2d pose0 = new Pose2d(0, 0, Math.toRadians(0));
-        Pose2d pose2 = new Pose2d(-3, 0, Math.toRadians(30));
-        Pose2d pose3 = new Pose2d(-26, 15, Math.toRadians(270));
-        Pose2d pose4 = new Pose2d(-26, 40, Math.toRadians(270));
-        Pose2d pose5 = new Pose2d(-3, 0, Math.toRadians(30));
-        Pose2d pose6 = new Pose2d(-50, 15, Math.toRadians(270));
-        Pose2d pose7 = new Pose2d(-50, 40, Math.toRadians(270));
-        Pose2d pose8 = new Pose2d(-3, 0, Math.toRadians(30));
-        MecanumDrive drive = new MecanumDrive(hardwareMap, pose0);
+        Pose2d initPose = new Pose2d(0, 0, Math.toRadians(0));
+        MecanumDrive drive = new MecanumDrive(hardwareMap, initPose);
         outtakeMotor shooter = new outtakeMotor(hardwareMap);
         transferServo transfer = new transferServo(hardwareMap);
         intakeServo intake = new intakeServo(hardwareMap);
         flickServo flicker = new flickServo(hardwareMap);
 
-        TrajectoryActionBuilder test = drive.actionBuilder(pose0)
-                .strafeToConstantHeading(new Vector2d(-3, 0), new TranslationalVelConstraint(50))
-                .turnTo(Math.toRadians(30))
-                .strafeToConstantHeading(new Vector2d(-26, -15), new TranslationalVelConstraint(50)) //counterclockwise by default
-                .turnTo(Math.toRadians(270))
-                .strafeToConstantHeading(new Vector2d(-26, -40), new TranslationalVelConstraint(15))
-                .strafeToConstantHeading(new Vector2d(-3, 0), new TranslationalVelConstraint(50))
-                .turnTo(Math.toRadians(30))
-                .strafeToConstantHeading(new Vector2d(-50, -15), new TranslationalVelConstraint(50))
-                .turnTo(Math.toRadians(270))
-                .strafeToConstantHeading(new Vector2d(-50, -40), new TranslationalVelConstraint(15))
-                .strafeToLinearHeading(new Vector2d(-3, 0), Math.toRadians(30), new TranslationalVelConstraint(50))
-                .strafeToConstantHeading(new Vector2d(-15, -15), new TranslationalVelConstraint(50));
+        TrajectoryActionBuilder one = drive.actionBuilder(initPose)
+                .turn(Math.toRadians(30)); //counterclockwise by default
 
-        TrajectoryActionBuilder one = drive.actionBuilder(pose0)
-                .strafeToConstantHeading(new Vector2d(-3, 0), new TranslationalVelConstraint(50))
-                .turnTo(Math.toRadians(30));
+        TrajectoryActionBuilder two = drive.actionBuilder(initPose)
+                .strafeToConstantHeading(new Vector2d(0, 22), new TranslationalVelConstraint(10))
+                .turn(Math.toRadians(125)) //counterclockwise by default
+                .strafeToConstantHeading(new Vector2d(0, 35), new TranslationalVelConstraint(10));
 
-        TrajectoryActionBuilder two = drive.actionBuilder(pose2)
-                .strafeToConstantHeading(new Vector2d(-26, -15), new TranslationalVelConstraint(10)) //counterclockwise by default
-                .turnTo(Math.toRadians(270))
-                .strafeToConstantHeading(new Vector2d(-26, -40), new TranslationalVelConstraint(10));
+        TrajectoryActionBuilder three = drive.actionBuilder(initPose)
+                .strafeToConstantHeading(new Vector2d(0, 10), new TranslationalVelConstraint(10));
 
+        TrajectoryActionBuilder four = drive.actionBuilder(initPose)
+                .strafeToConstantHeading(new Vector2d(0, -23), new TranslationalVelConstraint(10))
+                .turn(Math.toRadians(-125)) //counterclockwise by default
+                .strafeToConstantHeading(new Vector2d(0, -22), new TranslationalVelConstraint(10));
+        TrajectoryActionBuilder five = drive.actionBuilder(initPose)
+                .strafeToConstantHeading(new Vector2d(0, 12), new TranslationalVelConstraint(10));
 
-        TrajectoryActionBuilder three = drive.actionBuilder(pose3)
-                .strafeToConstantHeading(new Vector2d(-26, -40), new TranslationalVelConstraint(10));
-
-        TrajectoryActionBuilder four = drive.actionBuilder(pose4)
-                .strafeToConstantHeading(new Vector2d(-3, 0), new TranslationalVelConstraint(10))
-                .turnTo(Math.toRadians(30));
-
-        TrajectoryActionBuilder five = drive.actionBuilder(pose5)
-                .strafeToConstantHeading(new Vector2d(-50, -15), new TranslationalVelConstraint(10))
-                .turnTo(Math.toRadians(270));
-
-        TrajectoryActionBuilder six = drive.actionBuilder(pose6)
-                .strafeToConstantHeading(new Vector2d(-50, -40), new TranslationalVelConstraint(10));
-
-        TrajectoryActionBuilder seven = drive.actionBuilder(pose7)
-                .strafeToConstantHeading(new Vector2d(-3, 0), new TranslationalVelConstraint(10))
-                .turnTo(Math.toRadians(30));
-
-        TrajectoryActionBuilder eight = drive.actionBuilder(pose8)
-                .strafeToConstantHeading(new Vector2d(-15, -15), new TranslationalVelConstraint(10));
 
         // actions that need to happen on init
-
 
 
         waitForStart();
@@ -325,11 +295,11 @@ public class far_blue_9_no_mecha extends LinearOpMode{
 
         Actions.runBlocking(
                 new SequentialAction(
-                        //shooter.fireUp(),
-                        //shooter.hold(),
-                        //one.build(),
+                        shooter.fireUp(),
+                        shooter.hold(),
+                        one.build(),
 
-                        /*transfer.toA(),
+                        transfer.toA(),
                         flicker.kick(),
                         flicker.goBack(),
                         transfer.toB(),
@@ -337,19 +307,11 @@ public class far_blue_9_no_mecha extends LinearOpMode{
                         flicker.goBack(),
                         transfer.toC(),
                         flicker.kick(),
-                        flicker.goBack(),*/
+                        flicker.goBack(),
 
-                        test.build()
 
-                        //one.build(),
-                        //two.build(),
-                        //three.build(),
-                        //four.build(),
-                        //five.build(),
-                        //six.build(),
-                        //seven.build(),
-                        //eight.build()
-                        /*new ParallelAction( //TODO: the transfer timer should be longer for intaking than for outtaking
+                        two.build(),
+                        new ParallelAction( //TODO: the transfer timer should be longer for intaking than for outtaking
                                 intake.intaking(),
                                 three.build(),
                                 new SequentialAction(
@@ -372,7 +334,7 @@ public class far_blue_9_no_mecha extends LinearOpMode{
                         flicker.goBack(),
                         five.build(),
 
-                        shooter.stop()*/
+                        shooter.stop()
 
                 )
